moved around stuff + fixed dockerfile
