+++ cloudcompchem/exceptions.py
