+++ deploy/Dockerfile
@@ -1,18 +1,16 @@
-FROM python:3.8-alpine
+FROM python:3.11.8-bullseye
 
 # Add all the python code we need to run the app.
-RUN mkdir -p /opt/tachyon_runtime/
+RUN mkdir -p /opt/compchem_runtime/
 
 
-COPY ./tachyon/svc/ /opt/tachyon_runtime/
-COPY ./tachyon/constellation-sdk/ /opt/tachyon_runtime/
-COPY ./tachyon/requirements/common.txt /opt/tachyon_runtime/common.txt
-COPY ./tachyon/requirements/deployment.txt /opt/tachyon_runtime/requirements.txt
+ADD ./svc /opt/compchem_runtime
+ADD ./cloudcompchem /opt/compchem_runtime/cloudcompchem
+COPY ./setup.* /opt/compchem_runtime/
+COPY ./requirements/common.txt /opt/compchem_runtime/common.txt
+COPY ./requirements/deployment.txt /opt/compchem_runtime/requirements.txt
 
-WORKDIR /opt/tachyon_runtime
-
-# Upgrade pip to the latest version - this can fix some installation issues
-RUN python -m pip install -U pip
+WORKDIR /opt/compchem_runtime
 
 #################################################################
 ### Add any needed dependencies/libraries here.  Instructions ###
@@ -20,20 +18,28 @@ RUN python -m pip install -U pip
 ### https://wiki.alpinelinux.org/wiki/Package_management      ###
 #################################################################
 
-# Install bind utils
-RUN apk update && apk add bind-tools
+# security updates
+RUN DEBIAN_FRONTEND=noninteractive apt-get update && DEBIAN_FRONTEND=noninteractive apt-get upgrade -y
+
+# if we do need anything, append this...
+# DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+#    mypkg \
+#    && rm -rf /var/lib/apt/lists/*
+
+# Upgrade pip to the latest version - this can fix some installation issues
+RUN python -m pip install -U pip
 
 # Add ECL directory to pythonpath
 
-ENV PYTHONPATH "${PYTHONPATH}:/opt/tachyon_runtime/constellation-sdk"
+ENV PYTHONPATH "${PYTHONPATH}:/opt/compchem_runtime"
 
 # Install the requirements
 RUN pip install -r requirements.txt
 
 # Prepare the environment for Flask
 ENV FLASK_ENV production
-ENV FLASK_APP /opt/tachyon_runtime/svc/app.py
+ENV FLASK_APP /opt/compchem_runtime/svc/app.py
 
 # Run the flask app!
 EXPOSE 5000
-CMD gunicorn -b 0.0.0.0:5000 "wsgi:app" --log-level=INFO --timeout=90 --workers=8 --chdir /opt/tachyon_runtime/
+CMD gunicorn -b 0.0.0.0:5000 "wsgi:app" --log-level=INFO --timeout=90 --workers=8 --chdir /opt/compchem_runtime/
