+++ deploy/build.sh
@@ -14,10 +14,10 @@ echo "Finding the version..."
 GIT_COMMIT="$(git rev-parse HEAD)"
 ISODATE="$(date +%Y-%m-%dT%H:%M:%S%z)"
 DEPLOY_TAG="${ISODATE:0:10}--${GIT_COMMIT:0:8}"
-IMAGE_NAME="emeraldsci/tachyon:$DEPLOY_TAG"
+IMAGE_NAME="emeraldsci/cloudcompchem:$DEPLOY_TAG"
 
 echo "Building..."
-docker build --pull -f deploy/Dockerfile -t "$IMAGE_NAME" ..
+docker build --pull -f deploy/Dockerfile -t "$IMAGE_NAME" .
 docker push $IMAGE_NAME
 
 echo "Building $IMAGE_NAME complete!"
@@ -25,8 +25,7 @@ echo "Building $IMAGE_NAME complete!"
 echo
 echo "========================================================================="
 echo "== To deploy this, please email tachyon@emeraldcloudlab.com with your  =="
-echo "== image name and the constellation id of the simulation this image    =="
-echo "== supports.                                                           =="
+echo "== image name                                                          =="
 echo "========================================================================="
 echo
 echo "Image Name: $IMAGE_NAME"
