+++ requirements/common.txt
@@ -5,3 +5,6 @@ gunicorn==20.1.0
 backoff==1.10.0
 pysll==0.1.1
 pyscf==2.5.0
+
+# install cloudcompchem locally
+-e .
