+++ setup.cfg
@@ -1,0 +1,11 @@
+[metadata]
+name = cloudcompchem
+version = 0.1.0
+
+[options]
+packages = find:
+zip_safe = False
+
+[flake8]
+max-line-length = 120
+extend-ignore = E203
