+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup()
